pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "billable"
version = "0.1.0"
description = "Universal Billing Engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2",
    "django-ninja>=1.0",
    "Pillow>=10.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "billable.tests.test_settings"
pythonpath = ["."]
addopts = "-v"
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]              
include = ["billable*"]     